Update SentryLaravelServiceProvider.php
The current implementation assumes the default Laravel auth implementation. For anyone who uses auth schemes that aren't part of the core, this leaves lots of debris on the command line: `sentry.breadcrumbs error=Auth guard driver [api] is not defined.`

This change makes this optional behavior and doesn't alter it in any way.