Fix auth context not working for Laravel >=5.3
Because of changes in how the session is loaded in Laravel 5.3 we can no longer rely on getting the user inside the register method of the service provider. But there is an authenticated event introduced making it easy to simply listen for it and we can be sure the user is available.

Fixes #73