Add tag with name of console command
Listen for the `CommandStarting` event and add a tag with the name of the console command. This makes it easy to know which console command has the error.

Note: I did not typehint `ComandStarting` in `commandStartingHandler` because then I think it would only work on Laravel >=5.5 and based on the documentation this package supports Laravel 5.*. I'm surprised that the typehint of `MessageLogged` on  `messageLoggedHandler` does not cause issues for people trying to use older versions of Laravel since `MessageLogged` was not added until Laravel 5.4.