Fix repeatable booted callback registration
The `Middleware` class was relying on booted callback to measure the amount of time it takes for application to bootstrap, observing the difference between the Laravel start time (using the defined constant) or request start time from SAPI (if one is available) and falling back onto the timestamp when the request hit middleware. There were two issues with this approach:
- once request hits middleware, the framework has already finished booting by this point in time and the registered booted callback is fired immediately anyway, so there is no reason to have a step of indirection like that;
- but secondly, and more importantly, adding booted callback on each request meant that if the middleware to be used in long running process, such as when running from RoadRunner, Swoole or PSR-7 based web server and reusing application instances, they will accumulate and eventually tripping the memory limit or be OOM-killed;

therefore, after a brief conversation in #472, we agreed on inlining the booted callback into the middleware itself, which this commit is addressing.