fix (batch): check variable instead of index
In array_reduce the first arguments in the callback value holds the returned value from the last iteration. On initialization this first argument is equal to the third argument provided in the array_reduce instantiation. If no third argument is provided, as is in this case, the default value NULL is used.

Therefore on the first run $highest is equal to NULL and does not have the index 'level' available.

fixes getsentry/sentry-laravel#514
See https://www.php.net/manual/en/function.array-reduce.php