feat (fingers-crossed): enable action_level config
This update enables the user to provide the Sentry log channel with an action_level configuration. The inspiration for the configuration is taken from the Symfony Monolog configuration style.

All records at or above the provided level configuration will be buffered but not written to the logs. When a log record with a level at or above the action_level configuration is provided all the previously buffered log records will be flushed as context['logs'] and under Breadcrumbs in the Sentry log record.

closes getsentry/sentry-laravel#516
See https://symfony.com/doc/current/logging.html#handlers-that-modify-log-entries